Move shoebot.data to shoebot.graphics.
Some python files now reflect their class names better -
img to image
bezier to bezierpath
