+++ shoebot/graphics/basecolor.py
