+++ shoebot/graphics/bezierpath.py
