+++ shoebot/graphics/geometry.py
