+++ shoebot/graphics/image.py
