+++ shoebot/graphics/point.py
