+++ shoebot/graphics/variable.py
