+++ tests/unittests/test_text.py
@@ -6,7 +6,7 @@ from tests.unittests.helpers import ShoebotTestCase
 from textwrap import dedent
 
 from shoebot.core import CairoCanvas, CairoImageSink
-from shoebot.data import Text
+from shoebot.graphics import Text
 from shoebot.grammar import NodeBot
